Define MPAs UI overhauled to use generic item management engine - one we should have had ages ago
git-svn-id: https://sources.ecopath.org/svn/Ecopath/trunk/Ecopath6@16384 a725721c-0ed6-864e-a6a8-02c2a1988c9d

+++ Sources/ScientificInterfaceShared/Controls/EwEGrid/EwEGrid.vb
@@ -1504,10 +1504,10 @@ Namespace Controls.EwEGrid
         ''' selections. An assertion may occur otherwise.
         ''' </remarks>
         ''' -------------------------------------------------------------------
-        Public Sub SelectRow(ByVal iRow As Integer)
+        Public Sub SelectRow(ByVal iRow As Integer, Optional bAddToSelection As Boolean = False)
 
             ' Clear current selection
-            If (Me.Selection IsNot Nothing) Then
+            If (Me.Selection IsNot Nothing And Not bAddToSelection) Then
                 Dim r As SourceGrid2.Range = Me.Selection.GetRange()
                 If Not r.IsEmpty Then
                     Me.Selection.RemoveRange(r)
